+++ packages/liquid-js-wallet-provider/README.md
@@ -1,0 +1,11 @@
+# `liquid-js-wallet-provider`
+
+> TODO: description
+
+## Usage
+
+```
+const liquidJsWalletProvider = require('liquid-js-wallet-provider');
+
+// TODO: DEMONSTRATE API
+```
