+++ packages/liquid-js-wallet-provider/lib/index.js
@@ -1,0 +1,3 @@
+import LiquidJsWalletProvider from './LiquidJsWalletProvider'
+
+export default LiquidJsWalletProvider
