+++ packages/liquid-js-wallet-provider/package.json
@@ -1,17 +1,17 @@
 {
-  "name": "@liquality/liquid-wallet-provider",
+  "name": "@liquality/liquid-js-wallet-provider",
   "version": "0.0.1",
-  "description": "Blockstream Liquid Wallet Provider",
+  "main": "lib/index.js",
+  "description": "> TODO: description",
   "keywords": [
     "liquid",
-    "wallet",
+    "blockstream",
     "elements",
-    "blockstream"
+    "bitcoin"
   ],
   "author": "Marco Argentieri <3596602+tiero@users.noreply.github.com>",
   "homepage": "https://github.com/liquality/chainabstractionlayer#readme",
   "license": "MIT",
-  "main": "lib/index.js",
   "directories": {
     "lib": "lib",
     "test": "__tests__"
@@ -30,15 +30,9 @@
     "url": "https://github.com/liquality/chainabstractionlayer/issues"
   },
   "dependencies": {
-    "@liquality/utils": "^0.7.1",
-    "@liquality/wallet-provider": "^0.7.1",
     "bip32": "^2.0.6",
+    "bip39": "^3.0.2",
     "liquidjs-lib": "github:vulpemventures/liquidjs-lib",
     "slip77": "github:vulpemventures/slip77"
-  },
-  "devDependencies": {
-    "bip39": "^3.0.2",
-    "jest": "^26.5.3",
-    "mocha": "^8.1.3"
   }
 }
